Move effects to schemas and scripts folders
