+++ projects/Eevee/Eevee.csproj
@@ -38,10 +38,10 @@
     <Folder Include="souchy.celebi.eevee\face\shared\conditions\value\" />
     <Folder Include="souchy.celebi.eevee\impl\net\amethyst\" />
     <Folder Include="souchy.celebi.eevee\impl\shared\conditions\value\" />
-    <Folder Include="souchy.celebi.eevee\neweffects\impl\effects\move\translate\" />
-    <Folder Include="souchy.celebi.eevee\neweffects\impl\effects\spell\" />
-    <Folder Include="souchy.celebi.eevee\neweffects\impl\effects\status\" />
-    <Folder Include="souchy.celebi.eevee\neweffects\impl\effects\special\" />
+    <Folder Include="souchy.celebi.eevee\neweffects\impl\scripts\move\translate\" />
+    <Folder Include="souchy.celebi.eevee\neweffects\impl\scripts\spell\" />
+    <Folder Include="souchy.celebi.eevee\neweffects\impl\scripts\status\" />
+    <Folder Include="souchy.celebi.eevee\neweffects\impl\scripts\special\" />
   </ItemGroup>
 
   <ItemGroup>
