+++ projects/Eevee/souchy.celebi.eevee/neweffects/impl/schemas/SpellOnlySchemas.cs
@@ -7,7 +7,7 @@ using souchy.celebi.eevee.impl.objects.zones;
 using souchy.celebi.eevee.impl.stats;
 using souchy.celebi.eevee.neweffects.face;
 
-namespace souchy.celebi.eevee.neweffects.impl.effects
+namespace souchy.celebi.eevee.neweffects.impl.schemas
 {
 
     public interface SpellMetaSchema : IEffectSchema { }
