+++ projects/Eevee/souchy.celebi.eevee/neweffects/impl/scripts/move/teleport/TeleportTargetToScript.cs
@@ -3,30 +3,22 @@ using souchy.celebi.eevee.face.objects;
 using souchy.celebi.eevee.face.objects.effectResults;
 using souchy.celebi.eevee.impl.objects.effectReturn;
 using souchy.celebi.eevee.neweffects.face;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using souchy.celebi.eevee.neweffects.impl.schemas;
 
 namespace souchy.celebi.eevee.neweffects.impl.effects.move.teleport
 {
-    public class TeleportSelfToScript : IEffectScript
+    public class TeleportTargetToScript : IEffectScript
     {
-        public Type SchemaType => typeof(TeleportSelfTo);
+        public Type SchemaType => typeof(TeleportTargetTo);
 
         public IEffectReturnValue apply(ISubActionEffect action, IBoardEntity currentTarget, IEnumerable<IBoardEntity> allTargetsInZone)
         {
-            ICreature source = action.fight.creatures.Get(action.caster);
-            ICell cell = (ICell) currentTarget;
+            TeleportTargetTo props = action.effect.GetProperties<TeleportTargetTo>();
+            //props.MoveTargetZone
+            
 
-            var creatures = action.fight.board.GetCreaturesOnCell(cell.entityUid);
-            if (cell.type.isWalkable && creatures.Count() == 0)
-            {
-                source.position.set(cell.position);
-            }
 
-            return null;
+            throw new NotImplementedException();
         }
 
         public IEffectPreview preview(ISubActionEffect action, IBoardEntity currentTarget, IEnumerable<IBoardEntity> allTargetsInZone)
