+++ projects/Eevee/souchy.celebi.eevee/neweffects/impl/scripts/res/ADamageScript.cs
@@ -12,21 +12,22 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using souchy.celebi.eevee.neweffects.impl.util;
+using souchy.celebi.eevee.neweffects.impl.schemas;
 
 namespace souchy.celebi.eevee.neweffects.impl.effects.res
 {
-    public abstract class AHealScript : IEffectScript
+    public abstract class ADamageScript : IEffectScript
     {
         public abstract Type SchemaType { get; }
 
         protected abstract bool appliesOffensiveStats { get; }
         protected abstract bool appliesDefensiveStats { get; }
-
         public IEffectReturnValue apply(ISubActionEffect action, IBoardEntity currentTarget, IEnumerable<IBoardEntity> allTargetsInZone)
         {
             var creaSource = action.fight.creatures.Get(action.caster);
-            var creaTarget = (ICreature) currentTarget;
             var sourceStats = creaSource.GetTotalStats(action);
+
+            var creaTarget = (ICreature) currentTarget;
             var targetStats = creaTarget.GetTotalStats(action);
 
             var props = action.effect.GetProperties<ADamageSchema>();
@@ -35,18 +36,18 @@ namespace souchy.celebi.eevee.neweffects.impl.effects.res
             var shield = targetStats.Get<IStatSimple>(Resource.Shield);
             var remainingDmgAfterShield = dmg;
 
-            //var startShield = shield.value;
-            //shield.value -= dmg;
-            //shield.value = Math.Max(0, shield.value);
-            //remainingDmgAfterShield = startShield - shield.value;
+            var startShield = shield.value;
+            shield.value -= dmg;
+            shield.value = Math.Max(0, shield.value);
+            remainingDmgAfterShield = startShield - shield.value;
 
             var life = targetStats.Get<IStatSimple>(Resource.Life);
-            var lifeMax = targetStats.Get<IStatSimple>(Resource.LifeMax);
-            life.value += remainingDmgAfterShield;
+            life.value -= remainingDmgAfterShield;
 
             //var compiled = new EffectPreviewDamage(damage);
             //return compiled;
-            return new IEffectReturnValue(action.effect, dmg);
+            //return new IEffectReturnValue(e, dmg);
+            return null;
         }
 
         public IEffectPreview preview(ISubActionEffect action, IBoardEntity currentTarget, IEnumerable<IBoardEntity> allTargetsInZone)
@@ -54,6 +55,7 @@ namespace souchy.celebi.eevee.neweffects.impl.effects.res
             var props = action.effect.GetProperties<DirectDamage>();
             throw new NotImplementedException();
         }
+
 
     }
 }
