+++ projects/Eevee/souchy.celebi.eevee/neweffects/impl/scripts/res/TransferLifeScript.cs
@@ -11,6 +11,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using souchy.celebi.eevee.neweffects.impl.schemas;
 
 namespace souchy.celebi.eevee.neweffects.impl.effects.res
 {
